rsspy-backend/requirements.txt
amqp==5.0.9
asgiref==3.4.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cycler==0.11.0
DateTime==4.3
Django==4.0.1
django-cors-headers==3.10.1
django-crontab==0.7.1
django-extensions==3.1.5
django-scheduler==0.9.5
djangorestframework==3.13.1
feedparser==6.0.8
fonttools==4.28.5
icalendar==4.0.9
idna==3.3
kiwisolver==1.3.2
kombu==5.2.3
matplotlib==3.5.1
numpy==1.22.0
packaging==21.3
Pillow==9.0.0
prompt-toolkit==3.0.24
pycoingecko==2.2.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
schedule==1.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0
django-crontab==0.7.1
